Reworked the navigation bar and added another page for logout confirmation
