+++ Beer/Views/Components/HeaderComponents/Header.xaml
@@ -11,24 +11,84 @@
     xmlns:navcomponents="using:WinUIApp.Views.Components"
     mc:Ignorable="d">
 
-    <Grid Height="60" Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <StackPanel Orientation="Horizontal"
-                    VerticalAlignment="Center"
-                    HorizontalAlignment="Center">
-            <Button x:Name="GoBackButton"
-                    Visibility="{x:Bind GoBackButtonVisibility, Mode=OneWay}"
-                    Content="Go Back">
-            </Button>
-        </StackPanel>
+    <Grid Height="70">
+        <!-- Gradient Background -->
+        <Grid.Background>
+            <LinearGradientBrush StartPoint="0,0" EndPoint="1,0">
+                <GradientStop Color="#0F0F1A" Offset="0"/>
+                <GradientStop Color="#1A1A2E" Offset="0.3"/>
+                <GradientStop Color="#4B6CDB" Offset="0.7"/>
+                <GradientStop Color="#1A1A2E" Offset="1"/>
+            </LinearGradientBrush>
+        </Grid.Background>
+
+        <!-- Subtle bottom border -->
+        <Rectangle Height="1" VerticalAlignment="Bottom" Fill="#4BCDDB" Opacity="0.3"/>
+
+        <!-- Left-aligned Home Button -->
+        <Button x:Name="HomeButton"
+                Content="Home"
+                Click="GoHomeButton_Click"
+                Margin="30,0,0,0"
+                HorizontalAlignment="Left"
+                VerticalAlignment="Center"
+                Background="Transparent"
+                Foreground="White"
+                FontSize="14"
+                FontWeight="SemiBold"
+                Padding="12,6"
+                BorderThickness="0">
+            <Button.Resources>
+                <!-- Hover/Pressed states directly in the button -->
+                <Style TargetType="Button">
+                    <Setter Property="Template">
+                        <Setter.Value>
+                            <ControlTemplate TargetType="Button">
+                                <Grid>
+                                    <VisualStateManager.VisualStateGroups>
+                                        <VisualStateGroup x:Name="CommonStates">
+                                            <VisualState x:Name="Normal"/>
+                                            <VisualState x:Name="PointerOver">
+                                                <Storyboard>
+                                                    <ObjectAnimationUsingKeyFrames Storyboard.TargetName="ContentPresenter" Storyboard.TargetProperty="Foreground">
+                                                        <DiscreteObjectKeyFrame KeyTime="0" Value="#4BCDDB"/>
+                                                    </ObjectAnimationUsingKeyFrames>
+                                                </Storyboard>
+                                            </VisualState>
+                                            <VisualState x:Name="Pressed">
+                                                <Storyboard>
+                                                    <ObjectAnimationUsingKeyFrames Storyboard.TargetName="ContentPresenter" Storyboard.TargetProperty="Foreground">
+                                                        <DiscreteObjectKeyFrame KeyTime="0" Value="#DB4BCD"/>
+                                                    </ObjectAnimationUsingKeyFrames>
+                                                    <ObjectAnimationUsingKeyFrames Storyboard.TargetName="ContentPresenter" Storyboard.TargetProperty="Opacity">
+                                                        <DiscreteObjectKeyFrame KeyTime="0" Value="0.8"/>
+                                                    </ObjectAnimationUsingKeyFrames>
+                                                </Storyboard>
+                                            </VisualState>
+                                        </VisualStateGroup>
+                                    </VisualStateManager.VisualStateGroups>
+                                    <ContentPresenter x:Name="ContentPresenter"
+                                                    Content="{TemplateBinding Content}"
+                                                    HorizontalAlignment="Center"
+                                                    VerticalAlignment="Center"/>
+                                </Grid>
+                            </ControlTemplate>
+                        </Setter.Value>
+                    </Setter>
+                </Style>
+            </Button.Resources>
+        </Button>
+
+        <!-- Right-aligned controls -->
         <StackPanel Orientation="Horizontal" 
                     HorizontalAlignment="Right" 
                     VerticalAlignment="Center"
-                    Margin="0 0 30 0">
+                    Margin="0,0,30,0"
+                    Spacing="15">
             <components:AddDrinkButton x:Name="AddDrinkButtonControl" 
                                      VerticalAlignment="Center"/>
             <navcomponents:NavigationMenu x:Name="NavMenu"
-                                      VerticalAlignment="Center"
-                                      Margin="20 0 0 0"/>
+                                      VerticalAlignment="Center"/>
         </StackPanel>
     </Grid>
 </UserControl>
