+++ Beer/Views/Components/HeaderComponents/Header.xaml.cs
@@ -34,35 +34,9 @@ namespace WinUIApp.Views.Components.HeaderComponents
             }
         }
 
-        private void GoBackButton_Click(object sender, RoutedEventArgs routedEventArgs)
-        {
-            if (AuthenticationWindow.PreviousPage != null)
-            {
-                AuthenticationWindow.NavigationFrame.Navigate(AuthenticationWindow.PreviousPage);
-            }
-            else
-            {
-                AuthenticationWindow.NavigationFrame.Navigate(typeof(MainPage));
-            }
-        }
-
-        private void Logo_Click(object sender, RoutedEventArgs routedEventArgs)
+        private void GoHomeButton_Click(object sender, RoutedEventArgs routedEventArgs)
         {
             AuthenticationWindow.NavigationFrame.Navigate(typeof(MainPage));
-        }
-
-        public Visibility GoBackButtonVisibility { get; set; } = Visibility.Collapsed;
-
-        public void SetGoBackButtonVisibility(bool isVisible)
-        {
-            GoBackButtonVisibility = isVisible ? Visibility.Visible : Visibility.Collapsed;
-            // this.Bindings.Update();
-        }
-
-        public void UpdateHeaderComponentsVisibility(Type currentPageType)
-        {
-            bool shouldShowGoBackButton = currentPageType != typeof(MainPage);
-            SetGoBackButtonVisibility(shouldShowGoBackButton);
         }
     }
 }
