+++ Beer/Views/Components/NavigationMenu.xaml.cs
@@ -1,13 +1,11 @@
+using System;
+using DataAccess.Constants;
+using DataAccess.Service.Interfaces;
+using DrinkDb_Auth;
+using DrinkDb_Auth.View;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using DrinkDb_Auth;
-using DataAccess.Service.Interfaces;
-using Microsoft.Extensions.DependencyInjection;
-using DataAccess.Constants;
-using WinUIApp;
-using System;
-using DrinkDb_Auth.View;
-using WinUIApp.Views.Pages;
 
 namespace WinUIApp.Views.Components
 {
@@ -52,18 +50,7 @@ namespace WinUIApp.Views.Components
 
         private void LogoutButton_Click(object sender, RoutedEventArgs e)
         {
-            // Clear the current user ID and session
-            App.CurrentUserId = Guid.Empty;
-            App.CurrentSessionId = Guid.Empty;
-            
-            // Get the window from the XAML root
-            var window = (AuthenticationWindow)App.MainWindow;
-            window.HandleLogout();
-        }
-
-        private void HomeButton_Click(object sender, RoutedEventArgs e)
-        {
-            AuthenticationWindow.NavigationFrame.Navigate(typeof(MainPage));
+            AuthenticationWindow.NavigationFrame.Navigate(typeof(ConfirmLogoutPage));
         }
     }
-} 
+}
