+++ Beer/Views/Pages/ConfirmLogoutPage.xaml
@@ -1,0 +1,132 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="DrinkDb_Auth.ConfirmLogoutPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:DrinkDb_Auth"
+    xmlns:components="using:WinUIApp.Views.Components"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid>
+        <!-- Dark background for better contrast with gradient circles -->
+        <Grid.Background>
+            <LinearGradientBrush StartPoint="0,0" EndPoint="1,1">
+                <GradientStop Color="#0F0F1A" Offset="0"/>
+                <GradientStop Color="#1A1A2E" Offset="1"/>
+            </LinearGradientBrush>
+        </Grid.Background>
+
+        <!-- Animated Gradient Circles -->
+        <Grid>
+            <Grid.Resources>
+                <Storyboard x:Name="CircleAnimation" RepeatBehavior="Forever">
+                    <!-- Circle 1 - Moves diagonally -->
+                    <DoubleAnimationUsingKeyFrames Storyboard.TargetName="Circle1" Storyboard.TargetProperty="(UIElement.RenderTransform).(CompositeTransform.TranslateX)">
+                        <LinearDoubleKeyFrame KeyTime="0:0:0" Value="-200"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:20" Value="300"/>
+                    </DoubleAnimationUsingKeyFrames>
+                    <DoubleAnimationUsingKeyFrames Storyboard.TargetName="Circle1" Storyboard.TargetProperty="(UIElement.RenderTransform).(CompositeTransform.TranslateY)">
+                        <LinearDoubleKeyFrame KeyTime="0:0:0" Value="-100"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:20" Value="200"/>
+                    </DoubleAnimationUsingKeyFrames>
+
+                    <!-- Circle 2 - Moves horizontally with slight vertical variation -->
+                    <DoubleAnimationUsingKeyFrames Storyboard.TargetName="Circle2" Storyboard.TargetProperty="(UIElement.RenderTransform).(CompositeTransform.TranslateX)">
+                        <LinearDoubleKeyFrame KeyTime="0:0:0" Value="100"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:25" Value="-300"/>
+                    </DoubleAnimationUsingKeyFrames>
+                    <DoubleAnimationUsingKeyFrames Storyboard.TargetName="Circle2" Storyboard.TargetProperty="(UIElement.RenderTransform).(CompositeTransform.TranslateY)">
+                        <LinearDoubleKeyFrame KeyTime="0:0:0" Value="150"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:12.5" Value="50"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:25" Value="150"/>
+                    </DoubleAnimationUsingKeyFrames>
+
+                    <!-- Circle 3 - Moves vertically with slight horizontal variation -->
+                    <DoubleAnimationUsingKeyFrames Storyboard.TargetName="Circle3" Storyboard.TargetProperty="(UIElement.RenderTransform).(CompositeTransform.TranslateX)">
+                        <LinearDoubleKeyFrame KeyTime="0:0:0" Value="-100"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:15" Value="50"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:30" Value="-100"/>
+                    </DoubleAnimationUsingKeyFrames>
+                    <DoubleAnimationUsingKeyFrames Storyboard.TargetName="Circle3" Storyboard.TargetProperty="(UIElement.RenderTransform).(CompositeTransform.TranslateY)">
+                        <LinearDoubleKeyFrame KeyTime="0:0:0" Value="300"/>
+                        <LinearDoubleKeyFrame KeyTime="0:0:30" Value="-200"/>
+                    </DoubleAnimationUsingKeyFrames>
+                </Storyboard>
+            </Grid.Resources>
+
+            <!-- Gradient Circles -->
+            <Ellipse x:Name="Circle1" Width="450" Height="450" Opacity="0.25">
+                <Ellipse.Fill>
+                    <RadialGradientBrush>
+                        <GradientStop Color="#FF4B6CDB" Offset="0"/>
+                        <GradientStop Color="#FF4B6CDB" Offset="0.5"/>
+                        <GradientStop Color="#00FFFFFF" Offset="1"/>
+                    </RadialGradientBrush>
+                </Ellipse.Fill>
+                <Ellipse.RenderTransform>
+                    <CompositeTransform/>
+                </Ellipse.RenderTransform>
+            </Ellipse>
+
+            <Ellipse x:Name="Circle2" Width="350" Height="350" Opacity="0.25">
+                <Ellipse.Fill>
+                    <RadialGradientBrush>
+                        <GradientStop Color="#FFDB4BCD" Offset="0"/>
+                        <GradientStop Color="#FFDB4BCD" Offset="0.5"/>
+                        <GradientStop Color="#00FFFFFF" Offset="1"/>
+                    </RadialGradientBrush>
+                </Ellipse.Fill>
+                <Ellipse.RenderTransform>
+                    <CompositeTransform/>
+                </Ellipse.RenderTransform>
+            </Ellipse>
+
+            <Ellipse x:Name="Circle3" Width="500" Height="500" Opacity="0.25">
+                <Ellipse.Fill>
+                    <RadialGradientBrush>
+                        <GradientStop Color="#FF4BCDDB" Offset="0"/>
+                        <GradientStop Color="#FF4BCDDB" Offset="0.5"/>
+                        <GradientStop Color="#00FFFFFF" Offset="1"/>
+                    </RadialGradientBrush>
+                </Ellipse.Fill>
+                <Ellipse.RenderTransform>
+                    <CompositeTransform/>
+                </Ellipse.RenderTransform>
+            </Ellipse>
+        </Grid>
+
+        <!-- Content Container with slight blur effect -->
+        <Grid>
+            <Grid.Background>
+                <SolidColorBrush Color="#20000000" Opacity="0.7"/>
+            </Grid.Background>
+            <Border CornerRadius="16" Padding="40" Background="#801A1A2E">
+                <StackPanel VerticalAlignment="Center" HorizontalAlignment="Center" Spacing="20">
+                    <TextBlock Text="Are you sure you want to logout?"
+                             FontSize="32"
+                             FontWeight="Bold"
+                             HorizontalAlignment="Center"
+                             Foreground="White"/>
+
+                    <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Spacing="20">
+                        <Button x:Name="ContinueButton"
+                                Content="Continue"
+                                Click="ContinueButton_Click"
+                                Style="{ThemeResource AccentButtonStyle}"
+                                Padding="24,12"
+                                FontSize="16"/>
+
+                        <Button x:Name="CancelButton"
+                            Content="Cancel"
+                            Click="CancelButton_Click"
+                            Style="{ThemeResource AccentButtonStyle}"
+                            Padding="24,12"
+                            FontSize="16"/>
+                    </StackPanel>
+                </StackPanel>
+            </Border>
+        </Grid>
+    </Grid>
+</Page>
