+++ Beer/Views/Pages/ProfilePage.xaml
@@ -81,12 +81,6 @@
             <RowDefinition Height="*"/>
         </Grid.RowDefinitions>
 
-        <components:NavigationMenu x:Name="NavMenu" 
-                                 Grid.Row="0"
-                                 HorizontalAlignment="Right"
-                                 Margin="0,10,20,0"
-                                 IsAdmin="{x:Bind IsAdmin, Mode=OneWay}"/>
-
         <!-- Main content area -->
         <Grid Grid.Row="1">
             <!-- User Profile Section - Fixed position -->
